<threaddump> fix bug 921537 so ruby thread dump reports useful log file #cherrypick
https://bugzilla.redhat.com/show_bug.cgi?id=921537 rhc threaddump have wrong log file prompt
upstream is:
https://bugzilla.redhat.com/show_bug.cgi?id=923405 thread dumps report the wrong log file name due to timezone
cherry-picked origin-server commit e83866b1b1753bde0a47e33fb2a94e0eeb436e76

When ruby does a thread dump it goes in the httpd server log. This is
rotated via rotatelogs and the name includes the timezone in it. Due to
various complications this does not always match the TZ from `date`.
Rather than try to give the exact file name, we now just give a glob
that is good enough.