<perl/bin/build> Bug 971403 - Tweak build script to account for unset HOME variable
This change replaces a ~ with the V1 OPENSHIFT_HOMEDIR env var (which
remains in V2 carts) to prevent incorrect filespec when HOME is
unset. It also tests the HOME variable and if it's unset, exports it
with the value in OPENSHIFT_HOMEDIR. This prevents cpanm (which
doesn't know about OPENSHIFT_HOMEDIR) from trying to modify the root
directory.